reorganize the util functions away from the business logic
